backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pdfplumber
python-docx
scikit-learn
